+++ src/neural_network/layer/average_pooling_3d.rs
@@ -213,18 +213,5 @@ impl Layer for AveragePooling3D {
         "AveragePooling3D"
     }
 
-    fn output_shape(&self) -> String {
-        if !self.input_shape.is_empty() {
-            let output_shape =
-                calculate_output_shape_3d_pooling(&*self.input_shape, self.pool_size, self.strides);
-            format!(
-                "({}, {}, {}, {}, {})",
-                output_shape[0], output_shape[1], output_shape[2], output_shape[3], output_shape[4]
-            )
-        } else {
-            "unknown".to_string()
-        }
-    }
-
-    no_trainable_parameters_layer_functions!();
+    layer_functions_3d_pooling!();
 }