+++ src/neural_network/layer/global_average_pooling_2d.rs
@@ -157,12 +157,5 @@ impl Layer for GlobalAveragePooling2D {
         "GlobalAveragePooling2D"
     }
 
-    fn output_shape(&self) -> String {
-        if self.input_shape.is_empty() {
-            return "unknown".to_string();
-        }
-        format!("[{}, {}]", self.input_shape[0], self.input_shape[1])
-    }
-
-    no_trainable_parameters_layer_functions!();
+    layer_functions_avg_pooling!();
 }